Hack around CI permission errors
In the CI env, GOPATH is set to /go and GOROOT is set to /usr/local/go.
But the former has subdirectories -- notably /go/pkg -- that are not
writable by the CI user; so when we try to do `go tool`ish things, we
get permission errors.

Work around this by making /go recursively group writable, which it
really ought to be anyway.